+++ tools/restore.rb
@@ -1,0 +1,41 @@
+require_relative '../config/initialize'
+require_relative '../db/connect'
+Dir[File.join(APPLICATION_ROOT, 'models/*')].each {|f| require_relative f }
+
+logger = ZosmaLogger.new(Settings.logger.path.restore)
+ApplicationRecord.zosma_logger = logger
+
+begin
+  from = ARGV.find {|arg| arg.start_with?('--from=') }
+  from = from ? Date.parse(from.match(/\A--from=(.*)\z/)[1]) : (Date.today - 2)
+  to = ARGV.find {|arg| arg.start_with?('--to=') }
+  to = to ? Date.parse(to.match(/\A--to=(.*)\z/)[1]) : Date.today
+rescue ArgumentError => e
+  logger.error(e.backtrace.join("\n"))
+  raise e
+end
+
+logger.info('==== Start restore ====')
+logger.info("FROM: #{from.strftime('%F')}")
+logger.info("TO: #{to.strftime('%F')}")
+start_time = Time.now
+
+(from..to).each do |date|
+  [
+    ['rate', Rate],
+    ['candle_stick', CandleStick],
+    ['moving_average', MovingAverage],
+  ].each do |type, klass|
+    backup_dir = File.join(APPLICATION_ROOT, Settings.import.file[type].backup_dir)
+    file_name = File.join(backup_dir, "#{date.strftime('%F')}.csv")
+
+    unless File.exist?(file_name)
+      logger.warn("#{file_name} is not exist")
+      next
+    end
+
+    klass.load_data(file_name)
+  end
+end
+
+logger.info("==== Finish restore (run_time: #{Time.now - start_time}) ====")